requirements.txt
pbr>=0.5.21,<1.0
Babel>=0.9.6

extras
python-daemon
